requirements.txt
asgiref==3.3.4
config==0.5.0.post0
dj-database-url==0.5.0
Django==3.2.1
django-heroku==0.3.1
django-phonenumber-field==5.1.0
gunicorn==20.1.0
mysqlclient==2.0.3
phonenumbers==8.12.22
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
